feat - mod login security

+++ Backend/jobterview/src/main/java/ssafy/project/jobterview/repository/UserRepository.java
@@ -3,5 +3,6 @@ package ssafy.project.jobterview.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ssafy.project.jobterview.domain.Member;
 
-public interface UserRepository extends JpaRepository<Member, Integer> {
+public interface UserRepository extends JpaRepository<Member, Long> {
+
 }
